requirements.txt
google-generativeai
python-dotenv
streamlit
requests
pytest
